+++ playroom/snippets.tsx
@@ -584,6 +584,7 @@ const listRowSnippets: Array<Snippet> = ['Row', 'BoxedRow'].flatMap((rowName) =>
         title="Title"
         description="Description"
         iconButton={{
+          "aria-label": "Borrar",
           Icon: IconTrashCanRegular,
           small: true,
           backgroundType: "transparent",
@@ -955,6 +956,18 @@ const cardSnippets: Array<Snippet> = [
               onPress: () => {},
               label: "Lightning",
             },
+            {
+              checkedProps: {
+                  Icon: IconStarFilled,
+                  label: 'checked',
+              },
+              uncheckedProps: {
+                  Icon: IconStarRegular,
+                  label: 'unchecked',
+              },
+              defaultChecked: false,
+              onChange: () => {},
+          },
           ]}
         />`,
     },
@@ -980,6 +993,18 @@ const cardSnippets: Array<Snippet> = [
               onPress: () => {},
               label: "Lightning",
             },
+            {
+              checkedProps: {
+                  Icon: IconStarFilled,
+                  label: 'checked',
+              },
+              uncheckedProps: {
+                  Icon: IconStarRegular,
+                  label: 'unchecked',
+              },
+              defaultChecked: false,
+              onChange: () => {},
+          },
           ]}
         />`,
     },
@@ -1009,6 +1034,7 @@ const cardSnippets: Array<Snippet> = [
           headline={<Tag type="promo">Headline</Tag>}
           pretitle="Pretitle"
           title="Title"
+          subtitle="Subtitle"
           description="Description"
           backgroundImage="https://source.unsplash.com/900x900/?landscape"
           onClose={() => {}}
@@ -1019,6 +1045,18 @@ const cardSnippets: Array<Snippet> = [
               onPress: () => {},
               label: "Lightning",
             },
+            {
+              checkedProps: {
+                  Icon: IconStarFilled,
+                  label: 'checked',
+              },
+              uncheckedProps: {
+                  Icon: IconStarRegular,
+                  label: 'unchecked',
+              },
+              defaultChecked: false,
+              onChange: () => {},
+          },
           ]}
         />`,
     },
@@ -1030,6 +1068,7 @@ const cardSnippets: Array<Snippet> = [
           headline={<Tag type="promo">Headline</Tag>}
           pretitle="Pretitle"
           title="Title"
+          subtitle="Subtitle"
           description="Description"
           backgroundVideo="https://fr-cert1-es.mytelco.io/2O4-xBJqiMlAfLkseq8RkXs_mv2ACV7Hnt20HqXxNl-mK7KLI3M2dAw"
           poster="https://source.unsplash.com/900x900/?landscape"
@@ -1049,6 +1088,7 @@ const cardSnippets: Array<Snippet> = [
           headline={<Tag type="promo">Headline</Tag>}
           pretitle="Pretitle"
           title="Title"
+          subtitle="Subtitle"
           description="Description"
           isInverse
           onClose={() => {}}
@@ -1059,6 +1099,18 @@ const cardSnippets: Array<Snippet> = [
               onPress: () => {},
               label: "Lightning",
             },
+            {
+              checkedProps: {
+                  Icon: IconStarFilled,
+                  label: 'checked',
+              },
+              uncheckedProps: {
+                  Icon: IconStarRegular,
+                  label: 'unchecked',
+              },
+              defaultChecked: false,
+              onChange: () => {},
+          },
           ]}
         />`,
     },
@@ -1070,6 +1122,7 @@ const cardSnippets: Array<Snippet> = [
           headline={<Tag type="promo">Headline</Tag>}
           pretitle="Pretitle"
           title="Title"
+          subtitle="Subtitle"
           description="Description"
           backgroundColor={colors.promo}
           isInverse
@@ -1080,6 +1133,18 @@ const cardSnippets: Array<Snippet> = [
               Icon: IconLightningRegular,
               onPress: () => {},
               label: "Lightning",
+            },
+            {
+              checkedProps: {
+                  Icon: IconStarFilled,
+                  label: 'checked',
+              },
+              uncheckedProps: {
+                  Icon: IconStarRegular,
+                  label: 'unchecked',
+              },
+              defaultChecked: false,
+              onChange: () => {},
             },
           ]}
         />`,