+++ src/card.tsx
@@ -13,7 +13,7 @@ import * as styles from './card.css';
 import * as mediaStyles from './image.css';
 import {useTheme} from './hooks';
 import {sprinkles} from './sprinkles.css';
-import {IconButton} from './icon-button';
+import {InternalIconButton, InternalToggleIconButton} from './icon-button';
 import IconCloseRegular from './generated/mistica-icons/icon-close-regular';
 import IconPauseFilled from './generated/mistica-icons/icon-pause-filled';
 import IconPlayFilled from './generated/mistica-icons/icon-play-filled';
@@ -40,16 +40,30 @@ import type {
     TrackingEvent,
 } from './utils/types';
 
-export type CardAction = {
-    label: string;
+type BaseIconButtonAction = {
     Icon: React.FC<IconProps>;
+    label: string;
+};
+
+type IconButtonAction = BaseIconButtonAction &
+    ExclusifyUnion<
+        | {href: string; newTab?: boolean}
+        | {to: string; fullPageOnWebView?: boolean; replace?: boolean}
+        | {onPress: () => void}
+    >;
+
+type ToggleIconButtonAction = {
+    checkedProps: BaseIconButtonAction;
+    uncheckedProps: BaseIconButtonAction;
+    onChange?: (checked: boolean) => void | undefined | Promise<void>;
+    checked?: boolean;
+    defaultChecked?: boolean;
+};
+
+export type CardAction = {
     disabled?: boolean;
     trackingEvent?: TrackingEvent | ReadonlyArray<TrackingEvent>;
-} & ExclusifyUnion<
-    | {href: string; newTab?: boolean}
-    | {to: string; fullPageOnWebView?: boolean; replace?: boolean}
-    | {onPress: () => void}
->;
+} & ExclusifyUnion<IconButtonAction | ToggleIconButtonAction>;
 
 const useTopActions = (actions?: Array<CardAction | React.ReactElement>, onClose?: () => void) => {
     const {texts} = useTheme();
@@ -75,40 +89,49 @@ type CardActionsGroupProps = {
     type?: 'default' | 'inverse' | 'media';
 };
 
-export const TOP_ACTION_BUTTON_SIZE = 48;
-
-export const CardActionIconButton = ({Icon, label, ...restProps}: CardAction): JSX.Element => {
+export const CardActionIconButton = (props: CardAction): JSX.Element => {
     const type = React.useContext(CardActionTypeContext);
 
-    const iconColor = {
-        default: vars.colors.neutralHigh,
-        inverse: vars.colors.inverse,
-        media: '#000000',
-    };
-
-    const iconBackgroundStyle = {
-        default: styles.cardAction,
-        inverse: styles.cardActionInverse,
-        media: styles.cardActionMedia,
-    };
     return (
-        <IconButton
-            {...restProps}
-            aria-label={label}
-            size={TOP_ACTION_BUTTON_SIZE}
-            className={styles.cardActionIconButton}
-            style={{display: 'flex'}}
-        >
-            <div className={iconBackgroundStyle[type]}>
-                <Icon color={iconColor[type]} size={20} />
-            </div>
-        </IconButton>
+        <ThemeVariant isInverse={type === 'inverse'}>
+            {/** we render IconButton if Icon prop was passed. Otherwise, ToggleIconButton will be used */}
+            {props.Icon ? (
+                <InternalIconButton
+                    {...props}
+                    aria-label={props.label}
+                    small
+                    isOverMedia={type === 'media'}
+                    type="neutral"
+                    backgroundType="transparent"
+                    hasInteractiveAreaBleed
+                />
+            ) : (
+                <InternalToggleIconButton
+                    {...props}
+                    checkedProps={{
+                        ...props.checkedProps,
+                        'aria-label': props.checkedProps.label,
+                        type: type === 'media' ? 'neutral' : 'brand',
+                        backgroundType: 'solid',
+                    }}
+                    uncheckedProps={{
+                        ...props.uncheckedProps,
+                        'aria-label': props.uncheckedProps.label,
+                        type: 'neutral',
+                        backgroundType: 'transparent',
+                    }}
+                    small
+                    isOverMedia={type === 'media'}
+                    hasInteractiveAreaBleed
+                />
+            )}
+        </ThemeVariant>
     );
 };
 
 export const CardActionsGroup = ({
     actions,
-    padding = 8,
+    padding = 16,
     onClose,
     type = 'default',
 }: CardActionsGroupProps): JSX.Element => {
@@ -125,16 +148,16 @@ export const CardActionsGroup = ({
                     zIndex: 3, // needed because images has zIndex 1 and touchable overlay has zIndex 2
                 }}
             >
-                <div className={sprinkles({display: 'flex'})}>
+                <Inline space={16}>
                     {finalActions.map((action, index) => {
-                        if ('label' in action) {
+                        if ('Icon' in action || 'checkedProps' in action) {
                             // action is a CardAction object
                             return <CardActionIconButton key={index} {...action} />;
                         }
                         // action is a React.ReactElement
                         return action;
                     })}
-                </div>
+                </Inline>
             </div>
         </CardActionTypeContext.Provider>
     ) : (
@@ -300,20 +323,20 @@ const useVideoWithControls = (
     }
 
     const videoAction: CardAction = {
-        Icon: {
-            playing: CardActionPauseIcon,
-            loading: CardActionPauseIcon,
-            paused: CardActionPlayIcon,
-            loadingTimeout: isRunningAcceptanceTest() ? CardActionPauseIcon : CardActionSpinner,
-        }[videoStatus],
-        onPress: onVideoControlPress,
-        label: {
-            playing: texts.pauseIconButtonLabel,
-            loading: texts.pauseIconButtonLabel,
-            paused: texts.playIconButtonLabel,
-            loadingTimeout: '',
-        }[videoStatus],
+        uncheckedProps: {
+            Icon:
+                videoStatus === 'loadingTimeout' && !isRunningAcceptanceTest()
+                    ? CardActionSpinner
+                    : CardActionPauseIcon,
+            label: videoStatus === 'loadingTimeout' ? '' : texts.pauseIconButtonLabel,
+        },
+        checkedProps: {
+            Icon: CardActionPlayIcon,
+            label: texts.playIconButtonLabel,
+        },
+        onChange: onVideoControlPress,
         disabled: videoStatus === 'loadingTimeout',
+        checked: videoStatus === 'paused',
     };
 
     return {
@@ -794,15 +817,10 @@ export const DataCard = React.forwardRef<HTMLDivElement, DataCardProps>(
         },
         ref
     ) => {
-        const hasIcon = !!icon;
+        const hasIconOrHeadline = !!icon || !!headline;
         const isTouchable = touchableProps.href || touchableProps.to || touchableProps.onPress;
 
         const finalActions = useTopActions(actions, onClose);
-        const topActionsStylesWithoutIcon = {
-            marginRight: -16,
-            marginTop: -24,
-            width: TOP_ACTION_BUTTON_SIZE * finalActions.length,
-        } as const;
 
         return (
             <CardContainer
@@ -846,7 +864,14 @@ export const DataCard = React.forwardRef<HTMLDivElement, DataCardProps>(
                                     />
                                 </Stack>
                                 {/** Hack to avoid content from rendering on top of the top action buttons */}
-                                {!hasIcon && <div style={topActionsStylesWithoutIcon} />}
+                                {!hasIconOrHeadline && (
+                                    <div
+                                        style={applyCssVars({
+                                            [styles.vars.topActionsCount]: String(finalActions.length),
+                                        })}
+                                        className={styles.dataCardTopActionsWithoutIcon}
+                                    />
+                                )}
                             </Inline>
 
                             {extra && <div>{extra}</div>}
@@ -1274,6 +1299,8 @@ interface PosterCardBaseProps {
     pretitleLinesMax?: number;
     title?: string;
     titleLinesMax?: number;
+    subtitle?: string;
+    subtitleLinesMax?: number;
     description?: string;
     descriptionLinesMax?: number;
 }
@@ -1323,6 +1350,8 @@ export const PosterCard = React.forwardRef<HTMLDivElement, PosterCardProps>(
             pretitleLinesMax,
             title,
             titleLinesMax,
+            subtitle,
+            subtitleLinesMax,
             description,
             descriptionLinesMax,
             variant,
@@ -1448,7 +1477,7 @@ export const PosterCard = React.forwardRef<HTMLDivElement, PosterCardProps>(
                                     <Stack space={24}>
                                         <div>
                                             <Stack space={8}>
-                                                {(headline || pretitle || title) && (
+                                                {(headline || pretitle || title || subtitle) && (
                                                     <header>
                                                         <Stack space={16}>
                                                             {headline}
@@ -1477,6 +1506,16 @@ export const PosterCard = React.forwardRef<HTMLDivElement, PosterCardProps>(
                                                                 >
                                                                     {title}
                                                                 </Text>
+                                                                <Text2
+                                                                    forceMobileSizes
+                                                                    truncate={subtitleLinesMax}
+                                                                    as="div"
+                                                                    regular
+                                                                    textShadow={textShadow}
+                                                                    hyphens="auto"
+                                                                >
+                                                                    {subtitle}
+                                                                </Text2>
                                                             </Stack>
                                                         </Stack>
                                                     </header>
@@ -1489,6 +1528,11 @@ export const PosterCard = React.forwardRef<HTMLDivElement, PosterCardProps>(
                                                         regular
                                                         textShadow={textShadow}
                                                         hyphens="auto"
+                                                        color={
+                                                            withGradient
+                                                                ? vars.colors.textPrimary
+                                                                : vars.colors.textSecondary
+                                                        }
                                                     >
                                                         {description}
                                                     </Text2>