+++ src/sheet.css.ts
@@ -168,7 +168,7 @@ export const modalCloseButton = style([
         position: 'absolute',
         top: 0,
         right: 0,
-        padding: 16,
+        padding: 24,
     }),
     {
         '@media': {