+++ src/skins/movistar-legacy.tsx
@@ -68,7 +68,8 @@ export const getMovistarLegacySkin: GetKnownSkin = (variant) => {
             backgroundOverlay: applyAlpha(palette.grey6, 0.6),
             backgroundSkeleton: palette.grey2,
             backgroundSkeletonInverse: palette.movistarBlue55,
-            backgroundFeedbackBottom: palette.movistarBlue,
+            backgroundBrandTop: palette.movistarBlue,
+            backgroundBrandBottom: palette.movistarBlue,
             appBarBackground: palette.white,
             navigationBarBackground: palette.movistarBlue,
             skeletonWave: palette.grey2,
@@ -112,7 +113,7 @@ export const getMovistarLegacySkin: GetKnownSkin = (variant) => {
             textLinkSnackbar: palette.movistarBlue30,
             control: palette.grey3,
             controlActivated: palette.movistarBlue,
-            controlInverse: palette.white,
+            controlInverse: palette.movistarBlue20,
             controlActivatedInverse: palette.white,
             controlError: palette.pepper,
             barTrack: palette.grey3,
@@ -164,6 +165,18 @@ export const getMovistarLegacySkin: GetKnownSkin = (variant) => {
             textAppBar: palette.grey4,
             textAppBarSelected: palette.movistarBlue,
             customTabsBackground: palette.white,
+            tagTextPromo: palette.purple70,
+            tagTextActive: palette.movistarBlue,
+            tagTextInactive: palette.grey5,
+            tagTextSuccess: palette.movistarGreen70,
+            tagTextWarning: palette.egg80,
+            tagTextError: palette.pepper70,
+            tagBackgroundPromo: palette.purple10,
+            tagBackgroundActive: palette.movistarBlue10,
+            tagBackgroundInactive: palette.grey2,
+            tagBackgroundSuccess: palette.movistarGreen10,
+            tagBackgroundWarning: palette.egg10,
+            tagBackgroundError: palette.pepper10,
         },
         darkModeColors: {
             background: palette.darkModeBlack,
@@ -181,7 +194,8 @@ export const getMovistarLegacySkin: GetKnownSkin = (variant) => {
             backgroundOverlay: applyAlpha(palette.darkModeGrey, 0.8),
             backgroundSkeleton: palette.grey6,
             backgroundSkeletonInverse: palette.grey6,
-            backgroundFeedbackBottom: palette.darkModeBlack,
+            backgroundBrandTop: palette.darkModeBlack,
+            backgroundBrandBottom: palette.darkModeBlack,
             appBarBackground: palette.darkModeGrey,
             navigationBarBackground: palette.darkModeBlack,
             skeletonWave: palette.grey5,
@@ -277,6 +291,18 @@ export const getMovistarLegacySkin: GetKnownSkin = (variant) => {
             textAppBar: palette.grey5,
             textAppBarSelected: palette.grey2,
             customTabsBackground: palette.darkModeBlack,
+            tagTextPromo: palette.purple70,
+            tagTextActive: palette.movistarBlue,
+            tagTextInactive: palette.grey5,
+            tagTextSuccess: palette.movistarGreen70,
+            tagTextWarning: palette.egg80,
+            tagTextError: palette.pepper70,
+            tagBackgroundPromo: palette.grey6,
+            tagBackgroundActive: palette.grey6,
+            tagBackgroundInactive: palette.grey6,
+            tagBackgroundSuccess: palette.grey6,
+            tagBackgroundWarning: palette.grey6,
+            tagBackgroundError: palette.grey6,
         },
         borderRadii: {
             avatar: '50%',
@@ -318,7 +344,7 @@ export const getMovistarLegacySkin: GetKnownSkin = (variant) => {
         skin.colors.brandHigh = palette.movistarProminentBlueDark;
         skin.colors.backgroundBrand = palette.movistarProminentBlue;
         skin.colors.navigationBarBackground = palette.movistarProminentBlue;
-        skin.colors.backgroundFeedbackBottom = palette.movistarProminentBlue;
+        skin.colors.backgroundBrandBottom = palette.movistarProminentBlue;
         skin.colors.controlActivated = palette.movistarProminentBlue;
         skin.colors.dividerInverse = palette.movistarProminentBlueDark;
     }